refactor: split up tests
